fix: Resolve f != java.lang.Long type mismatch - Convert all monetary fields from Double to Long (cents) CRITICAL BUG FIX: Root cause of 'f != java.lang.Long' error when saving invoices The error was caused by payment-related entities using Double for monetary amounts while the core Invoice system uses Long (cents). This type mismatch prevented the database from accepting saved invoice data. Changes Made: - InvoicePaymentEntity.amountPaid: Double → Long (cents) - InvoicePaymentSnapshot: All monetary fields Double → Long - DailyPaymentSnapshot: All monetary fields Double → Long - CollectionMetrics: All monetary fields Double → Long - PaymentAnalyticsRepository.recordPayment: amountPaid parameter Double → Long - PaymentAnalyticsRepositoryImpl: Updated to convert Long to Double in domain mapping Database Migration: - Created Migration_23_24.kt: Converts all monetary columns from REAL to INTEGER - Multiplies all existing values by 100 (e.g., 149.99 → 14999 cents) - Recreates all indices after column type changes - AppDatabase version: 23 → 24 Testing: - Build successful: 39s, APK 24.8 MB - All type mismatches resolved - Ready for runtime testing This fix ensures all monetary values use consistent Long (cents) representation throughout the persistence layer, eliminating the Float-to-Long casting error.

+++ Bizap/app/src/main/java/com/emul8r/bizap/data/local/entities/InvoicePaymentEntity.kt
@@ -24,7 +24,7 @@ data class InvoicePaymentEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
     val invoiceId: Long,
-    val amountPaid: Double,
+    val amountPaid: Long,  // Cents (e.g., 14999 = $149.99)
     val paymentDate: Long,  // Unix timestamp
     val paymentMethod: String,  // "CASH", "CHECK", "CREDIT_CARD", etc.
     val transactionReference: String,
@@ -70,17 +70,17 @@ data class InvoicePaymentSnapshot(
     val invoiceDate: Long,
     val dueDate: Long,
     
-    val totalAmount: Double,
-    val paidAmount: Double,
-    val outstandingAmount: Double,
-    
+    val totalAmount: Long,           // Cents
+    val paidAmount: Long,            // Cents
+    val outstandingAmount: Long,     // Cents
+
     val paymentStatus: String,
     val ageingBucket: String,
     val daysOverdue: Int,
     val daysSinceDue: Int,
     
     val lastPaymentDate: Long? = null,
-    val lastPaymentAmount: Double = 0.0,
+    val lastPaymentAmount: Long = 0,  // Cents
     val paymentCount: Int = 0,
     
     val isAtRisk: Boolean = false,
@@ -108,21 +108,21 @@ data class DailyPaymentSnapshot(
     val snapshotDate: Long,
     
     val paymentsReceivedCount: Int,
-    val paymentsReceivedAmount: Double,
+    val paymentsReceivedAmount: Long,      // Cents
     val invoicesDueCount: Int,
-    val invoicesDueAmount: Double,
+    val invoicesDueAmount: Long,           // Cents
     val invoicesOverdueCount: Int,
-    val invoicesOverdueAmount: Double,
-    
-    val outstandingCurrent: Double,
-    val outstandingPast30: Double,
-    val outstandingPast60: Double,
-    val outstandingPast90: Double,
-    
+    val invoicesOverdueAmount: Long,       // Cents
+
+    val outstandingCurrent: Long,          // Cents
+    val outstandingPast30: Long,           // Cents
+    val outstandingPast60: Long,           // Cents
+    val outstandingPast90: Long,           // Cents
+
     val collectionRate: Double,
     val averagePaymentTime: Double,
-    val projectedMonthlyRevenue: Double,
-    
+    val projectedMonthlyRevenue: Long,     // Cents
+
     val createdAtMs: Long = System.currentTimeMillis()
 )
 
@@ -136,26 +136,26 @@ data class CollectionMetrics(
     val metricsDate: Long,
     
     val totalInvoicesIssued: Int,
-    val totalInvoiceAmount: Double,
-    val totalPaidAmount: Double,
-    val totalOutstandingAmount: Double,
+    val totalInvoiceAmount: Long,      // Cents
+    val totalPaidAmount: Long,         // Cents
+    val totalOutstandingAmount: Long,  // Cents
     val collectionRate: Double,
     
-    val ageingCurrent: Double,
-    val ageingPast30: Double,
-    val ageingPast60: Double,
-    val ageingPast90: Double,
-    
+    val ageingCurrent: Long,           // Cents
+    val ageingPast30: Long,            // Cents
+    val ageingPast60: Long,            // Cents
+    val ageingPast90: Long,            // Cents
+
     val averageDaysToPayment: Double,
     val medianDaysToPayment: Double,
     val overdueInvoiceCount: Int,
-    val overdueAmount: Double,
-    
+    val overdueAmount: Long,           // Cents
+
     val collectionRateTrend: Double,
     val overdueTrend: Double,
     
     val projectedCollectionRate30Days: Double,
-    val projectedOutstanding30Days: Double,
-    
+    val projectedOutstanding30Days: Long,  // Cents
+
     val lastUpdatedMs: Long = System.currentTimeMillis()
 )
